feat: Add non-cached denormalize path
