+++ packages/normalizr/src/denormalize.ts
@@ -1,267 +1,26 @@
-import GlobalCache from './cache.js';
-import type { Schema, EntityInterface, UnvisitFunction } from './interface.js';
-import { isEntity } from './isEntity.js';
-import { denormalize as arrayDenormalize } from './schemas/Array.js';
-import { isImmutable } from './schemas/ImmutableUtils.js';
-import { denormalize as objectDenormalize } from './schemas/Object.js';
-import type {
-  Denormalize,
-  DenormalizeNullable,
-  DenormalizeCache,
-  Path,
-} from './types.js';
-import WeakEntityMap, {
-  type Dep,
-  getEntities,
-  type GetEntity,
-  depToPaths,
-} from './WeakEntityMap.js';
+import type { Schema } from './interface.js';
+import LocalCache from './localCache.js';
+import type { Denormalize, DenormalizeNullable } from './types.js';
+import getUnvisit from './unvisit.js';
+import { getEntities } from './WeakEntityMap.js';
 
-const unvisitEntity = (
-  entityOrId: Record<string, any> | string,
-  schema: EntityInterface,
-  unvisit: UnvisitFunction,
-  getEntity: GetEntity,
-  cache: GlobalCache,
-): [denormalized: object | undefined, found: boolean, deleted: boolean] => {
-  const entity =
-    typeof entityOrId === 'object'
-      ? entityOrId
-      : getEntity({ pk: entityOrId, key: schema.key });
-  if (
-    typeof entity === 'symbol' &&
-    (entity as symbol).toString().includes('DELETED')
-  ) {
-    return [undefined, true, true];
-    // TODO: Change to this as breaking change once we only support newer entities
-    // also remove `(entity as symbol).toString().includes('DELETED')` and do this for all symbols
-    // return schema.denormalize(entity, unvisit);
-  }
-
-  if (typeof entity !== 'object' || entity === null) {
-    return [entity as any, false, false];
-  }
-
-  const pk =
-    // normalize must always place a string, because pk() return value is string | undefined
-    // therefore no need to check for numbers
-    typeof entityOrId === 'string'
-      ? entityOrId
-      : schema.pk(isImmutable(entity) ? (entity as any).toJS() : entity);
-  // if we can't generate a working pk; this is hopeless so let's give them what's already there
-  // otherwise, even when we aren't doing a lookup we want to turn the entityOrId object into the
-  // expected class, and cache that class for referential equality. PK is used for global equality lookup.
-  if (pk === undefined || pk === '' || pk === 'undefined') {
-    return [entity, false, false];
-  }
-
-  // last function computes if it is not in any caches
-  return cache.get(pk, schema, entity, localCacheKey =>
-    setLocalCache(entity, schema, unvisit, pk, localCacheKey),
-  );
-};
-
-function setLocalCache(
-  entity: object,
-  schema: EntityInterface<any>,
-  unvisit: UnvisitFunction,
-  pk: string,
-  localCacheKey: Record<string, any>,
-): [found: boolean, deleted: boolean] {
-  let entityCopy: any, found, deleted;
-  if (schema.createIfValid) {
-    entityCopy = localCacheKey[pk] = isImmutable(entity)
-      ? schema.createIfValid(entity.toObject())
-      : schema.createIfValid(entity);
-    // TODO(breaking): remove once old verions no longer supported
-  } else {
-    entityCopy = entity;
-    unvisit = withTrackedEntities(unvisit);
-    unvisit.setLocal = entityCopy => (localCacheKey[pk] = entityCopy);
-  }
-
-  if (entityCopy === undefined) {
-    // undefined indicates we should suspense (perhaps failed validation)
-    found = false;
-    deleted = true;
-  } else {
-    [localCacheKey[pk], found, deleted] = schema.denormalize(
-      entityCopy,
-      unvisit,
-    );
-  }
-  return [found, deleted];
-}
-
-const getUnvisit = (
-  entities: Record<string, Record<string, any>>,
-  entityCache: DenormalizeCache['entities'],
-  resultCache: DenormalizeCache['results'][string],
-) => {
-  const getEntity = getEntities(entities);
-  const cache = new GlobalCache(getEntity, getEntityCaches(entityCache));
-
-  function unvisit(
-    input: any,
-    schema: any,
-  ): [denormalized: any, found: boolean, deleted: boolean] {
-    if (!schema) return [input, true, false];
-
-    // null is considered intentional, thus always 'found' as true
-    if (input === null) {
-      return [input, true, false];
-    }
-
-    const hasDenormalize = typeof schema.denormalize === 'function';
-
-    // deserialize fields (like Date)
-    if (!hasDenormalize && typeof schema === 'function') {
-      if (input instanceof schema) return [input, true, false];
-      // field deserialization should never count against 'found' (whether to used inferred results)
-      if (input === undefined) return [input, true, false];
-      return [new schema(input), true, false];
-    }
-
-    if (input === undefined) {
-      return [input, false, false];
-    }
-
-    if (!hasDenormalize && typeof schema === 'object') {
-      const method = Array.isArray(schema)
-        ? arrayDenormalize
-        : objectDenormalize;
-      return method(schema, input, unvisit);
-    }
-
-    if (isEntity(schema)) {
-      return unvisitEntity(input, schema, unvisit, getEntity, cache);
-    }
-
-    if (hasDenormalize) {
-      return schema.denormalize(input, unvisit);
-    }
-
-    return [input, true, false];
-  }
-
-  return (
-    input: any,
-    schema: any,
-  ): [
-    denormalized: any,
-    found: boolean,
-    deleted: boolean,
-    entityPaths: Path[],
-  ] => {
-    // in the case where WeakMap cannot be used
-    // this test ensures null is properly excluded from WeakMap
-    const cachable = Object(input) === input && Object(schema) === schema;
-    if (!cachable) {
-      const ret = unvisit(input, schema);
-      // this is faster than spread
-      // https://www.measurethat.net/Benchmarks/Show/23636/0/spread-with-tuples
-      return [ret[0], ret[1], ret[2], cache.paths()];
-    }
-
-    let [ret, entityPaths] = resultCache.get(input, getEntity);
-
-    if (ret === undefined) {
-      ret = unvisit(input, schema);
-      // we want to do this before we add our 'input' entry
-      entityPaths = cache.paths();
-      // for the first entry, `path` is ignored so empty members is fine
-      cache.dependencies.unshift({ entity: input, path: { key: '', pk: '' } });
-      resultCache.set(cache.dependencies, ret);
-    }
-
-    return [ret[0], ret[1], ret[2], entityPaths as Path[]];
-  };
-};
-
-const getEntityCaches = (entityCache: DenormalizeCache['entities']) => {
-  return (pk: string, schema: EntityInterface) => {
-    const key = schema.key;
-
-    if (!(key in entityCache)) {
-      entityCache[key] = Object.create(null);
-    }
-    const entityCacheKey = entityCache[key];
-    if (!entityCacheKey[pk])
-      entityCacheKey[pk] = new WeakMap<
-        EntityInterface,
-        WeakEntityMap<object, any>
-      >();
-
-    let wem: WeakEntityMap<object, any> = entityCacheKey[pk].get(schema) as any;
-    if (!wem) {
-      wem = new WeakEntityMap<object, any>();
-      entityCacheKey[pk].set(schema, wem);
-    }
-
-    return wem;
-  };
-};
-
-type DenormalizeReturn<S extends Schema> =
-  | [
-      denormalized: Denormalize<S>,
-      found: true,
-      deleted: false,
-      entityPaths: Path[],
-    ]
-  | [
-      denormalized: DenormalizeNullable<S>,
-      found: boolean,
-      deleted: true,
-      entityPaths: Path[],
-    ]
-  | [
-      denormalized: DenormalizeNullable<S>,
-      found: false,
-      deleted: boolean,
-      entityPaths: Path[],
-    ];
-
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const denormalize = <S extends Schema>(
-  input: unknown,
-  schema: S | undefined,
-  entities: any,
-  entityCache: DenormalizeCache['entities'] = {},
-  resultCache: DenormalizeCache['results'][string] = new WeakEntityMap(),
-): DenormalizeReturn<S> => {
-  // undefined mean don't do anything
-  if (schema === undefined) {
-    return [input, true, false, []] as [any, boolean, boolean, any];
-  }
-  if (input === undefined) {
-    return [undefined, false, false, []] as [any, boolean, boolean, any];
-  }
-  return getUnvisit(entities, entityCache, resultCache)(input, schema);
-};
-
-export const denormalizeSimple = <S extends Schema>(
   input: any,
   schema: S | undefined,
   entities: any,
-  entityCache?: DenormalizeCache['entities'],
-  resultCache?: DenormalizeCache['results'][string],
 ):
   | [denormalized: Denormalize<S>, found: true, deleted: false]
   | [denormalized: DenormalizeNullable<S>, found: boolean, deleted: true]
-  | [denormalized: DenormalizeNullable<S>, found: false, deleted: boolean] =>
-  denormalize(input, schema, entities, entityCache, resultCache).slice(
-    0,
-    3,
-  ) as any;
+  | [denormalized: DenormalizeNullable<S>, found: false, deleted: boolean] => {
+  // undefined means don't do anything
+  if (schema === undefined) {
+    return [input, true, false] as [any, boolean, boolean];
+  }
+  if (input === undefined) {
+    return [undefined, false, false] as [any, boolean, boolean];
+  }
+  const getEntity = getEntities(entities);
 
-// TODO(breaking): remove once unused
-function withTrackedEntities(unvisit: UnvisitFunction): UnvisitFunction {
-  // every time we nest, we want to unwrap back to the top.
-  // this is due to only needed the next level of nested entities for lookup
-  const originalUnvisit = unvisit.og || unvisit;
-  const wrappedUnvisit = (input: any, schema: any) =>
-    originalUnvisit(input, schema);
-  wrappedUnvisit.og = unvisit;
-  return wrappedUnvisit;
-}
+  const ret = getUnvisit(getEntity, new LocalCache(getEntity))(input, schema);
+  return [ret[0], ret[1], ret[2]];
+};
